requirements.txt
appdirs==1.4.4
asgiref==3.5.0
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.4
Django==4.0.3
django-cors-headers==3.12.0
django-crontab==0.7.1
django-environ==0.8.1
django-filter==21.1
djangorestframework==3.13.1
drf-yasg==1.20.0
filelock==3.6.0
factory-boy==3.2.1
Faker==13.11.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
packaging==21.3
phonenumbers==8.12.47
platformdirs==2.5.1
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.8
pytest==7.1.2
pytest-django==4.5.2
pytest-factoryboy==2.2.1
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.2.0
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.14.1
